requirements.txt
numpy~=1.23.1
keras~=2.9.0
pandas~=1.4.3
matplotlib~=3.5.2
opencv-python~=4.6.0.66
swifter~=1.2.0
Pillow~=9.2.0
httplib2~=0.14.0
Flask~=2.1.3
tensorflow~=2.9.1
faiss-cpu~=1.7.2
Flask-Cors~=3.0.10